Cleanup: move theme/accent manager stuff out of StatusBar class
kanging neobuddy89 and SpiritCroc idea

https://github.com/crdroidandroid/android_frameworks_base/commit/f9db9b5c1a48fceaf5c212d2970c9282ca625466

https://github.com/AICP/frameworks_base/commit/cc8deab922ec383ab4d551013645ddf28f0ce181

From: Alex Cruz <alex@dirtyunicorns.com>
Date: Mon, 24 Sep 2018 18:57:46 -0500
Subject: Move our ThemeAccentsUtils class to the core/java/com/android/internal/statusbar

With this move, we can call it from not only the SystemUI but other packages. This will
reduce the amount of code in packages like DU-Tweaks, Updater and Settings.

From: Alex Cruz <alex@dirtyunicorns.com>
Date: Mon, 24 Sep 2018 19:21:59 -0500
Subject: Fruity Pebbles

In the commit below we added the ability to change between Dark/Light and wallpaper
based themes.

d5de41e

Since then we've extended this to Settings, DU-Tweaks, Changelog and a few other
apps in system. We've added more styles and have gone all in on this theme thing.

One thing that was missing was the ability to change the system accent. While the
pixel accent was good enough, many users had asked us to add in the ability to
change the accents because they wanted something different. Several ROMs have gone
about this same thing in their own way but we didn't. We wanted to use what we already
had and just add on to it.

This commit mirrors what maxwen did but with the accent, so thanks to him for that!

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>